Macros have __repr__. For user-defined macros, this includes the definition.
State gains set() and get(), and __setitem__() and __getitem__() are redefined
in terms of these. __setitem__() loses its use_global param, which is now in set().
This param now actually works.

State gains __len__() [for the number of blocks] and __contains__().

State.add_state() renamed to add_block() because it was confusing.

Tests updated.