When we're scanning for a Number but we see see the name of a count register, we take the value of that register.
This does for Number what commit 0f1957 did for Dimen. However,
that commit solved a crash, whereas this commit doesn't. The
difference is that this bug caused Numbers to contain Numbers.
Since Numbers can be coerced to int, this wasn't visible;
it was just a bit more inefficient. However, attempting to
coerce a [Number which contained a Number] to int crashed,
because __int__() must return an int.

Regression test added.