Reduced use of the_object_itself param. This is a breaking change. Setting params can be undone by State.
Formerly, many get() and __getitem__() methods returned the value
of the object requested, unless you passed in the_object_itself=True,
when you got the object itself. Many of these methods no longer
accept t_o_i, and where they do, it defaults to False. If you
want the value, call .value on the returned object.

When a group is active, State records changes in order to reverse
them when the group ends. Before this commit, this only applied to
changes made by State itself. So if you wrote

   state['count0'] = 100

then that would be undone when the group ended. But if you wrote

   state['count0'].value = 100

it wouldn't, even though otherwise it's identical. Now RegisterTables
will tell their State about changes made to their members.
As part of this, RegisterTables take a state param in their constructors.