C_UserDefined uses the main Tokeniser for expansion
Before, it used a new Tokeniser based on NullSource.
I don't know why that seemed like a good idea.
It meant that if a macro was used in the definition of another macro,
and it had too few arguments, the inner macro couldn't pick up
from after the outer one was done. (Currying, in other words.)