Tokeniser never exhausts.
If you reach the end of a file, we return None indefinitely.
If you spin on None for more than a few turns, with no change
to pushback, we raise an error.

Value.optional_string() removed, because it does the same as
Tokeniser.optional_string().

Debug statements adjusted.

Tests updated.