Control.set_from_tokens, which deliberately always raises an error
If we can set a value for the control, then we will have overridden the method.
But people might try to set values for other controls, so we
give them a reasonable error.

Token.macro() cut; it seemed like a good idea a month ago but it
never worked and was never used.