Number, Dimen, Glue, and Muglue can be compared as appropriate.
Number to Number, Number to int, and Dimen to Dimen, work for < <= == != > and >=.

Glue to Glue works for == and !=. Muglue likewise.

The various _get_* in test/__init__.py can all take a "raw" parameter
to return the actual object and not its value.

_tokenise_and_get in test/__init__.py now knows that Tokenisers
return None for EOF, rather than exhausting like they used to.

Tests added and updated.