State's getindex and setindex allow integer suffixes to the names of controls
You could already things up in register tables like this:

	s['count12']

But that only worked for register tables. This is now extended so that
you can add an integer to the end of any control. State will look up
that control, then call the control's __getitem__() or __setitem__()
as appropriate.

This will be used to find font dimensions.