Tokens can be initialised without a category parameter
If you do, then the category is chosen according to the rules
on p213 of the TeXbook:

"In all of the cases listed so far, \the produces a result
that is a sequence of ASCII character tokens. Category code 12
(“other”) is assigned to each token, except that character code 32
gets category 10 (“space”). The same rule is used to assign category
codes to the tokens produced by \number, \romannumeral, \string,
\meaning, \jobname, and \fontname."