Rethink the way Expander works
Redesigned Expander's interface for the benefit of callers, rather than
to match the way Expander works internally.

Updated most calls to Expander, which as you can imagine touches almost
everything. Some tests still break.

This was a huge commit which was supposed to go into a branch for a PR,
but didn't. Sorry about that.