When State.__setitem__() finds a register, it doesn't check for controls with the same name.
This was caused by forgetting to break out of a loop over (registers, controls).
It never came to light until now because there was never a control with the same name
as a register prefix. We now have \box.

It didn't affect __getitem__() because that returns as soon as it finds a result.

This caused test_box_registers() to fail. It now passes.