\noexpand passes its argument out correctly
It used to return its argument, because we used to have a rule that
the result of handlers got yielded by the Expander. We got rid of
that rule because only a few handlers used it, and it was confusing.
This means that Expander now special-cases Noexpand.

Tests updated. This reverts commit 9eab7776, which deleted a test
for "\norun_code", which doesn't exist. This was a search-and-replace
error when we replaced expand() with run_code(), which slipped through.