Handle caret sequences in the middle of control names.
For example, \d^^6fg for \dog.

Tokeniser._handle_caret() is now an ordinary method rather than a
generator. It pushes a Token onto the incoming queue, rather than
yielding it itself. This means that the caller can decide what to
do about it. We need that in order for the control name parser to
use it.

Tests updated.