Parameters and tracing filters are now subtypes of C_ControlWord. (Tracing filters are filters for logs.)
This simplifies the code a fair amount.

The "_mode" and "_currentfont" pseudo-parameters are now handled in-class
by State. "_currentfont" is renamed to "_font".

Also, fixed a bug where some Modes would attempt to exercise the page builder,
which was only implemented in Vertical. It's now a no-op in all the others.

Closes issue #5.

Tests updated.