Document.add_box(), Document.save()
Previously, \shipout would send boxes one by one to the output driver, which
was accessible via doc['_output'].

Now, Document hangs on to the boxes until they're ready to be saved.
The boxes are kept in Document.output and you can add to them with Document.add_box().
doc['_output'] is therefore abolished.

Document.save() will create an output driver as appropriate, and
pass all the boxes on to it.

Tests updated, but new tests are still needed.