When an undelimited argument has been read, don't look at the next token yet.
An undelimited argument is found by looking for a "single"-- that is, either
one token or a series of tokens balanced by brackets. If there was only one
token, we would set single_grouping=-1 as a flag. We would then go round
the loop again, read the next token at the top, then check single_grouping==-1,
push the token we'd just read back onto the stack, and return. This made
the code a bit more straightforward.

However, reading the next token might involve reading two newlines-- especially
if the call we were dealing with was followed by a comment. If the
macro we're calling is declared "outer", its arguments can't contain
paragraph breaks. When it saw the two newlines, the check for paragraph breaks
would trigger wrongly.

Right, I think I deserve a cup of coffee now.