All controls are stored in doc[] with keys beginning with a backslash.
Previously, controls were stored under the bare name-- so, \if was doc['if'].
But that caused a namespace collision between control names of one letter
and active characters.

This also means that there's no risk of TeX code defining a control
with the same name as a system element, like doc['_font'].

In addition, Token gains a property "identifier". For control names and active characters,
this is the key to use to store it in Document. For all other categories,
it raises NotImplementedError.

Huge checkin.

Closes issue #28.