Call stack is correct even when a macro calls another macro at the start of the definition
We keep track of the read position using the "location" field of a Token,
which records where in the file that Token was read. The call stack is built from these locations.

However, sometimes a macro calls another macro right at the start of its definition.
For example:

```
   \def\a{Hello}
   \def\b{\a world!}
   \b
```

In this case, no tokens of \b have been seen when we start running \a,
so the location in the call frame for the call to \a will not have been updated from
the call to \b.

The solution is to store a "starts_at" field for every macro definition,
which updates the location in the same way a Token does.

Also, Tokeniser.error_position() is now visible in Expander too.

Test added.