Non-letter characters are also added to WordBoxes
This enables ligatures such as "``".

The rules are as given on p282 of the TeXbook.
However, that also specifies that \noboundary should be added to the boxes,
which isn't possible yet because we haven't implemented it.

Also, added a small amount of work towards fixing \showbox (issue #52).

Also, boxes with no contents have width/height/depth of 0pt and not 0 the integer.

Tests added and updated.