HBoxes contain Breakpoints, and this is visible in their "contents" property
Previously, we filtered out the Breakpoints on the fly, and there was a proxy class
which made them visible. This became unwieldy. Henceforth, if you put stuff in
an HBox, you should expect it to add Breakpoints.

Boxes now keep their contents in "contents", not "_contents", since
there's no reason to filter it anywhere.

The old HBox.contents property used to add a breakpoint at the end of the result.
Since we're now just returning the contents as we have them, this is no longer added.
If you want one, add it yourself.

Removed dependency on wrapt.

Tests updated.