doc['_output'] is now doc['_contents']; doc['_output'] is now the output driver.
This brings the actual code in line with the documentation, and makes things
less confusing. It also means that a Document can know its output driver
before it gets going, so Special can work. (It doesn't work yet.)

Also, Output.render() now takes no params; it gets what data it needs from
the Document it was passed at its creation.

The input streams still believe themselves to be called doc['_input'], but
that's for a later commit.

Tests updated.