Partial rewrite of the HTML generator
It works now. It didn't work well before.

The ideas that we used to use for responsive paras we now use for everything.
We set this up by turning the list of non-glue and glue that we get from the Document
into a list of WrittenWords, possibly with glue around them.

Tests added. They all pass. Further tests needed for integration.
Some docstring stuff needed. I will do it soon, but England just won the European Cup
and I'm exhausted.

Ref issue #44; this goes a long way towards closing that but
not all the way. It's good enough to merge to main now, though.