Implementation of \expandafter that doesn't work
Test added. Ref issue #73.

We currently don't have a way to tell when the second token is done expanding
but not done executing. For example, if t2 is a macro with three arguments,
we'd read the name of the macro, then all three arguments (which could be#
arbitrarily long), then we'd push t1, then we'd expand the macro named in t2.

Fixing this is going to overlap enough with issue #49 that we'll do them together.