Box.from_tokens() takes no inside_mode parameter
It's only implemented in the superclass, but it behaves differently in the superclass to the subclasses.

Box.from_tokens() parses a box of any type, such as "\hbox{...}"

In the subclasses, e.g. HBox.from_tokens(), we assume the \hbox part
has already been read. This parallels the old behaviour, where the
mode used to be supplied as a param to from_tokens(). Instead it's
read from an "inside_mode" property on the subclass.