Inner modes don't pass their result to their enclosing mode
For example,

	I like \tokenlist1=\hbox{cheese} and onions

will not pass the box containing "cheese" to the outer HBox;
it will push it to the stack, where the \tokenlist will find it
and assign it to its variable.

Previously, it could happen that the result was pushed to the
stack *and* passed to the outer box, resulting in the same object
existing twice in the list of the outer box's mode! This was
the problem behind the problem that commit 5ed227 fixed: calculating
a mode's result twice was not just inefficient but resulted in
any duplicate HBox getting re-wrapped.

Regression test added.