\hrule and \vrule don't execute potential parts of their spec
This led to a problem where there'd be a \dimen containing a length,
but the \dimen control was executed, making it think someone was
trying to set the value.

This was a good excuse to refactor that function, which was a mess.

Regression test added. Also general tests, since there were none.