Expander(single=<bool>) -> Expander(bounded=<str>) where <str> can be "no", "single", or "balanced".
If bounded is not 'no', then on_eof must be 'exhaust';
this was always important, but we're now checking for it.

yex.control.Discretionary rewritten and much simplified because of this.

Expander's bounded and on_eof params are checked to ensure
they're one of the permitted keywords. This wasn't done on
on_eof before, but it should have been.

Expander.single_shot() removed; issue #37 refers.

Tests added and updated.