remember_restore() checks whether to dereference based on class
If you pass a register or a parameter into r_r(), you want to restore
its value rather than the object itself. Previously we would check
whether the object had a "value" member and use that if so. But that
ran into trouble with Dimen, which uses "value" for its integer width
in sp.