Registers and register tables are now controls, of type C_Register and C_Array respectively.
"is_array" and "is_queryable" flags added to controls. "is_queryable" is to be used
instead of checking for the existence of "value" or "get_value".

yex.decorator.array() added.

Added the on_query() decorator method of the results of yex.decorator.control().

Rewritten yex.value.Value.unsigned_number() as get_value_from_tokens(),
so it's more straightforward and understandable.

Tests updated. Not all of them pass yet. This will be a solution to issue #6 when it works.