Adding VBoxes inside VBoxes causes them to be flattened.
For example, if we have a VBox containing [one, two, three],
and we add another VBox [four, five], it would previously result in

	[one, two, three, [four, five]]

Now it produces

	[one, two, three, four, five]

This means that we don't get lots of nested VBoxes which don't
change the output, which makes testing easier. It also brings
us in line with TeX, per \showbox.

This same pattern is true for:

	- any VBox added to a VBox
	- any VBox added to the list of a vertical mode
	- any VBox added to doc.output (even if it's added
		by a vertical mode, as with \shipout).

Tests updated.