Corrected routine to decode TFM files. Font.em and Font.ex properties added.
The existing code got confused if a width was negative, which is represented
in two's complement. It flipped the bits but omitted to increment.

In addition, the TFM routine now produces a result in scaled points directly,
rather than creating a float which then gets turned into a scaled points value;
this lost precision.

Tests added and updated.