Dimen.__repr__() shows infinite values even if they're from non-point classes
For example, an infinite Muglue might be 1fi. Previously, a bug
meant that in such a case it would have been displayed in the base unit, as 1mu,
as if it had been finite.

Values whose display unit has a unit size of 1 (currently only Muglue)
are displayed as integers (without printing the ".0").