Document accepts a "style" parameter. This can be set using the -B or --bare param to yex/__main__.
Controls no longer hold a document reference. There's no need.

The default list of values for catcodes is in the style, not
hardcoded in the Catcode array.

The "ch", "identifier", and "name" attributes of Controls have
been made more like similarly-named attributes of other Tokens.
This has had knockon effects around the tests.