run_code gains auto_save parameter, find='items', and 'saw_all' in the result
Previously, find='ch' etc reported all tokens seen. This worked
counterintuitively: if the outermost mode is vertical, and it saw
a Letter or Other token, it pushes it back and switches the
mode to horizontal. So we see that token twice. This means, for example,
that call='hello' returned 'hhello'.

Therefore the 'saw' field in the standard result
now filters out tokens seen while the outermost mode is in effect.
'saw_all' has the behaviour of the old 'saw'. You can also add '_all' to the end of
several other potential values of find, and it will look in
'saw_all' instead of 'saw'.