Leaders track whether they're vertical.
Creating boxes can close two groups afterwards, if we're two groups deeper
than when we started the box. This allows us to get rid of the
different "flavours" of groups, which were confusing and led to
too many special cases.

This has caused huge knockon effects in the tests, and led to
the creation of find='..._all' parameters in run_code();
see commit 979c82 for more details.

Various fixes to paragraph wrapping; in particular, we don't try to
adjust \topskip that isn't there!

\csname properly sets the new control to \relax if it doesn't exist.