Our own logging fixture sets the root logging priority to CRITICAL.
This is to avoid spamming the user. I have renamed the fixture accordingly,
although it doesn't matter much what it's called because it has auto_use=True.

Also, the check for the logfile looks for yex.log rather than yex.control.log.
This must have been a search-and-replace error over a year ago, and
I don't know why it was never a problem!