Back out the Queryable change; it was unworkable
The idea of using a subclass is that automated documentation
and runtime code would both be able to understand it.
But it meant that @Control.on_query would have had to
cast its target to Queryable and write it back into the
package object, which was just getting too baroque.